consul-svc-healthy: use system jq if available
If we have a system jq, for example on some images jq is already installed, use
that one instead of trying to use the one from the consul-scripts volume because
that one will only work with Alpine/musl based images.
Ror example the docker-edgex-mongo image needs to use this script but has Ubuntu
as a base image which is glibc based and thus incompatible with the jq we have
staged in the consul-scripts volume. So for that image, we should instead use jq
from the base image.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>