fix[security-services]: Address race condition for mongodb initialization
This fix addresses a race condition for mongodb initialization
wherein the secret store is set up, but has not generated
per-sevice Vault tokens, nor committed mongodb passwords to
the secret store.  This fix introduces a sentinel file that
is used to indicate secretstore-setup completion without
giving access to the secretstore to an untrusted component (consul),
AND also removes the dependence on a sensitive file (resp-init.json).

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>