fix[core,security-services]: Consul health check
The Dockerfile was updated in several ways. First, the source consul
image version was incremented from 1.7.0 to 1.7.2. The creation of
default consul directories (/consul/scripts & /consul/config) are now
created in a single RUN line. There are new "edgex" directories to house
scripts & config files under /edgex that are filled with assets from the
root build directory. This creates a cleaner image that is only
initialized with config and script files that are relevant at run time.

Brings me to my next point - an additional entrypoint script was
prepended to the existing entrypoint script (that is inherited from the
consul image). This new entrypoint script (edgex-consul-entrypoint.sh)
handles the initialization (movement) of config & script files to the
default consul directories based on a few new environment variables that
show up in the docker-compose files. If there are no environment
variables, there are defaults set for the variables and default files
are copied over at run time to start the consul image up. All command
flags sent to this entrypoint are sent along to the default consul image
entrypoint script and are processed as normal. The Dockerfile injects
the new entrypoint script and the Dockerfile points to the new script as
its default entrypoint.

The Make file was simply cleaned up and I removed the "clean", "test",
and "run" options - given they were either blank or invalid.

Lastly, there is a new directory called "config" that contains all of
the base JSON configuration files that are utilized to setup the checks
for in consul. I also added a new one to check redis. The files are
named with 2 digits at the beginning of the file name so that we have
*some* control over the order in which they are loaded in to consul.
Double zero ("00") gets loaded first and sets the stage to load all of
the other configurations.

Signed-off-by: Beau Frusetta <beau.frusetta@intel.com>